logaggregator: move syslog stuff to shared packages
Move RFC5424 parsing and Message struct into pkg/syslog/rfc5424, and
move RFC6587 split func to pkg/syslog/rfc6587.Split
