+++ logaggregator/ring/ring_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	. "github.com/flynn/flynn/Godeps/_workspace/src/github.com/flynn/go-check"
-	"github.com/flynn/flynn/logaggregator/rfc5424"
+	"github.com/flynn/flynn/pkg/syslog/rfc5424"
 )
 
 // Hook gocheck up to the "go test" runner
