+++ pkg/syslog/rfc5424/message.go
